release 4.1.1
I had to pull 4.1.0 because the bower build did not include
Traverson 3.1.0 as advertised but only 3.0.0 because I forgot to
npm install after bumping the version in package.json. :-(

Since I did not want different releases on npm and bower, I
npm unpublished 4.1.0 and removed the bower release by deleting
the tag.